Fix type mismatch in `cvxcore` file `src/ProblemData.hpp`

+++ cvxpy/cvxcore/src/ProblemData.hpp
@@ -87,7 +87,7 @@ public:
   /**
    * Returns the row index vector I as a contiguous 1D numpy array.
    */
-  void getI(double *values, int num_values) {
+  void getI(int *values, int num_values) {
     std::vector<int> I = TensorI[param_id][vec_idx];
     for (int i = 0; i < num_values; i++) {
       values[i] = I[i];
@@ -97,7 +97,7 @@ public:
   /**
    * Returns the column index vector J as a contiguous 1D numpy array.
    */
-  void getJ(double *values, int num_values) {
+  void getJ(int *values, int num_values) {
     std::vector<int> J = TensorJ[param_id][vec_idx];
     for (int i = 0; i < num_values; i++) {
       values[i] = J[i];
